+++ src/hkimjp/wil2/view.clj
@@ -6,20 +6,19 @@
 
 (def version "0.2.1-SNAPSHOT")
 
-(def menu "text-xl font-medium text-white px-1 hover:bg-sky-400")
+(def ^:private menu "text-xl font-medium text-white px-1 hover:bg-sky-400")
 
 (defn navbar []
   [:div.flex.bg-green-900.items-baseline.gap-x-8
    [:div.text-2xl.font-medium.text-white "WIL2"]
-   [:div.text-xl.font-medium.text-white
-    [:span.px-1.hover:bg-sky-400 "days"]]
-   [:div {:class menu} "my"]
-   [:div {:class menu} "list"]
-   [:div.text-xl.font-medium.text-white
-    [:form {:method "post" :action "/logout"}
-     (h/raw (anti-forgery-field))
-     [:button.px-1.text-white.hover:bg-sky-400 "logout"]]]
-   [:div.text-xl.font-medium.text-white "HELP"]])
+   [:div {:class menu} [:a {:href "/wil2/todays"} "todays"]]
+   [:div {:class menu} [:a {:href "/wil2/my"} "my"]]
+   [:div {:class menu} [:a {:href "/wil2/weeks"} "weeks"]]
+   [:form {:method "post" :action "/logout"}
+    (h/raw (anti-forgery-field))
+    [:button {:class menu} "logout"]]
+   [:div {:class menu} [:a {:href "/help"} "HELP"]]
+   [:div {:class menu} [:a {:href "/admin"} "admin"]]])
 
 (def footer
   [:div.text-base
@@ -45,8 +44,8 @@
    [:body {:hx-boost "true"}
     [:div
      (navbar)
-     [:div content]
-     [:div footer]]]])
+     content
+     footer]]])
 
 (defn page
   [content]