+++ CHANGELOG.md
@@ -2,11 +2,17 @@
 
 - no effect reload/reload
 - datascript
+- gitignored resources/public/assets/css/output.css?
 
 # 0.2.1-SNAPSHOT (2025-09-04)
 
 - `tailwindcss --watch=always` is a go.
-
+- added `hkimjp/datascript-storage-javatime` in dependencies.
+- navbar menu - todays, my, weeks, logout, help
+- added `hkimjp/wil2/my.clj`
+- added `hkimjp/wil2/todays.clj`
+- added `hkimjp/wil2/weeks.clj`
+- changed routes.clj - :as -> :refer [...]
 
 # 0.2.0 (2025-09-03)